refactor: simplify contact screens
