+++ packages/apps/crm/src/components/organisms/Contact/ContactBottom/ContactBottom.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import EditButton from '../../EditButton/EditButton';
+
+const ContactBottom = ({navigation, idContact}) => {
+  return (
+    <EditButton
+      onPress={() =>
+        navigation.navigate('ContactFormScreen', {
+          idContact: idContact,
+        })
+      }
+    />
+  );
+};
+
+export default ContactBottom;
