+++ packages/apps/crm/src/components/organisms/Contact/index.js
@@ -1,0 +1,4 @@
+export {default as ContactBottom} from './ContactBottom/ContactBottom';
+export {default as ContactDropdownCards} from './ContactDropdownCards/ContactDropdownCards';
+export {default as ContactHeader} from './ContactHeader/ContactHeader';
+export {default as ContactPartnerCard} from './ContactPartnerCard/ContactPartnerCard';
