+++ packages/apps/crm/src/screens/contact/ContactDetailsScreen.js
@@ -1,21 +1,28 @@
 import React, {useEffect} from 'react';
 import {ScrollView} from 'react-native';
-import {Screen, HeaderContainer} from '@axelor/aos-mobile-ui';
+import {Screen, HeaderContainer, NotesCard} from '@axelor/aos-mobile-ui';
 import {
   useSelector,
   HeaderOptionsMenu,
   useDispatch,
+  useTranslator,
 } from '@axelor/aos-mobile-core';
-import {ContactBody, ContactBottom, ContactHeader} from '../../components';
-import {fetchContactEventById} from '../../features/eventSlice';
-import {fetchPartner} from '../../features/partnerSlice';
+import {
+  ContactBottom,
+  ContactDropdownCards,
+  ContactHeader,
+  ContactPartnerCard,
+} from '../../components';
 import {getContact} from '../../features/contactSlice';
 
 const ContactDetailsScreen = ({navigation, route}) => {
   const idContact = route.params.idContact;
+  const I18n = useTranslator();
   const dispatch = useDispatch();
+
   const {mobileSettings} = useSelector(state => state.config);
   const {contact} = useSelector(state => state.contact);
+
   React.useLayoutEffect(() => {
     navigation.setOptions({
       headerRight: () => (
@@ -34,15 +41,6 @@ const ContactDetailsScreen = ({navigation, route}) => {
     dispatch(getContact({contactId: idContact}));
   }, [dispatch, idContact]);
 
-  useEffect(() => {
-    dispatch(fetchContactEventById(idContact));
-  }, [dispatch, idContact]);
-
-  useEffect(() => {
-    contact?.mainPartner &&
-      dispatch(fetchPartner({partnerId: contact?.mainPartner?.id}));
-  }, [dispatch, contact?.mainPartner?.id, contact?.mainPartner]);
-
   return (
     <Screen removeSpaceOnTop={true}>
       <HeaderContainer
@@ -50,7 +48,9 @@ const ContactDetailsScreen = ({navigation, route}) => {
         fixedItems={<ContactHeader />}
       />
       <ScrollView>
-        <ContactBody navigation={navigation} />
+        <ContactPartnerCard navigation={navigation} />
+        <NotesCard title={I18n.t('Crm_Notes')} data={contact.description} />
+        <ContactDropdownCards navigation={navigation} />
       </ScrollView>
       <ContactBottom idContact={idContact} navigation={navigation} />
     </Screen>
